+++ packages/file-storage-convex/CHANGELOG.md
@@ -1,5 +1,7 @@
 # @vexcms/file-storage-convex
 
+## 0.1.0-alpha.16
+
 ## 0.1.0-alpha.15
 
 ## 0.1.0-alpha.14