+++ packages/richtext-plate/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @vexcms/richtext
 
+## 0.1.0-alpha.16
+
+### Patch Changes
+
+- Updated dependencies
+  - @vexcms/core@0.1.0-alpha.16
+
 ## 0.1.0-alpha.15
 
 ### Patch Changes